+++ bridge/Symfony/Normalizer/ScalarArrayDenormalizer.php
@@ -6,8 +6,7 @@ namespace Steevanb\PhpTypedArray\Bridge\Symfony\Normalizer;
 
 use Steevanb\PhpTypedArray\{
     AbstractTypedArray,
-    ScalarArray\AbstractScalarArray,
-    ScalarArray\ScalarArray
+    ScalarArray\ScalarArrayInterface
 };
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 
@@ -15,27 +14,19 @@ class ScalarArrayDenormalizer implements DenormalizerInterface
 {
     public function supportsDenormalization($data, $type, $format = null): bool
     {
-        return
-            $type === ScalarArray::class
-            || is_subclass_of($type, ScalarArray::class)
-            || $type === AbstractScalarArray::class
-            || is_subclass_of($type, AbstractScalarArray::class);
+        $interfaces = class_implements($type);
+
+        return is_array($interfaces) && array_key_exists(ScalarArrayInterface::class, $interfaces);
     }
 
-    /**
-     * Because of ScalarArray who extends AbstractTypedArray wen can't return AbstractScalarArray
-     *
-     * @var array<mixed> $data
-     * @var array<mixed> $context
-     */
-    public function denormalize($data, $type, $format = null, array $context = []): AbstractTypedArray
+    /** @var array<mixed> $context */
+    public function denormalize(mixed $data, string $type, $format = null, array $context = []): AbstractTypedArray
     {
         return $this
             ->createScalarArray($type)
             ->setValues($data);
     }
 
-    /** Because of ScalarArray who extends AbstractTypedArray wen can't return AbstractScalarArray */
     protected function createScalarArray(string $type): AbstractTypedArray
     {
         return new $type();