+++ tests/Unit/AbstractEnumArray/TestAbstractEnumArray.php
@@ -8,6 +8,11 @@ use Steevanb\PhpTypedArray\EnumArray\AbstractEnumArray;
 
 class TestAbstractEnumArray extends AbstractEnumArray
 {
+    public function __construct(string $className = TestEnum::class, iterable $values = [])
+    {
+        parent::__construct($className, $values);
+    }
+
     public function callCastValueToString(\UnitEnum $enum): string
     {
         return $this->castValueToString($enum);