+++ tests/Unit/ScalarArray/StringNullableArrayTest.php
@@ -8,15 +8,15 @@ use PHPUnit\Framework\TestCase;
 use Steevanb\PhpTypedArray\{
     Exception\InvalidTypeException,
     Exception\ValueAlreadyExistException,
-    ScalarArray\StringArray,
+    ScalarArray\StringNullableArray,
     ValueAlreadyExistsModeEnum
 };
 
-final class StringArrayTest extends TestCase
+final class StringNullableArrayTest extends TestCase
 {
     public function testAllowString(): void
     {
-        $array = new StringArray(['4']);
+        $array = new StringNullableArray(['4']);
 
         static::assertCount(1, $array);
         static::assertSame('4', $array[0]);
@@ -26,34 +26,35 @@ final class StringArrayTest extends TestCase
     {
         static::expectException(InvalidTypeException::class);
         /** @phpstan-ignore-next-line */
-        new StringArray([1]);
+        new StringNullableArray([1]);
     }
 
     public function testInvalidTypeFloat(): void
     {
         static::expectException(InvalidTypeException::class);
         /** @phpstan-ignore-next-line */
-        new StringArray([3.1]);
+        new StringNullableArray([3.1]);
     }
 
     public function testInvalidTypeBool(): void
     {
         static::expectException(InvalidTypeException::class);
         /** @phpstan-ignore-next-line */
-        new StringArray([true]);
+        new StringNullableArray([true]);
     }
 
-    public function testInvalidTypeNull(): void
+    public function testAllowNull(): void
     {
-        static::expectException(InvalidTypeException::class);
-        /** @phpstan-ignore-next-line */
-        new StringArray([null]);
+        $array = new StringNullableArray([null]);
+
+        static::assertCount(1, $array);
+        static::assertNull($array[0]);
     }
 
     public function testMergeValueAlreadyExistsAdd(): void
     {
-        $array = (new StringArray(['foo', 'bar'], ValueAlreadyExistsModeEnum::ADD))
-            ->merge(new StringArray(['bar', 'baz']));
+        $array = (new StringNullableArray(['foo', 'bar'], ValueAlreadyExistsModeEnum::ADD))
+            ->merge(new StringNullableArray(['bar', 'baz']));
 
         static::assertCount(4, $array);
         static::assertSame('foo', $array[0]);
@@ -64,8 +65,8 @@ final class StringArrayTest extends TestCase
 
     public function testMergeValueAlreadyExistsDoNotAdd(): void
     {
-        $array = (new StringArray(['foo', 'bar'], ValueAlreadyExistsModeEnum::DO_NOT_ADD))
-            ->merge(new StringArray(['bar', 'baz']));
+        $array = (new StringNullableArray(['foo', 'bar'], ValueAlreadyExistsModeEnum::DO_NOT_ADD))
+            ->merge(new StringNullableArray(['bar', 'baz']));
 
         static::assertCount(3, $array);
         static::assertSame('foo', $array[0]);
@@ -77,7 +78,7 @@ final class StringArrayTest extends TestCase
     public function testMergeValueAlreadyExistsException(): void
     {
         static::expectException(ValueAlreadyExistException::class);
-        (new StringArray(['foo', 'bar'], ValueAlreadyExistsModeEnum::EXCEPTION))
-            ->merge(new StringArray(['bar', 'baz']));
+        (new StringNullableArray(['foo', 'bar'], ValueAlreadyExistsModeEnum::EXCEPTION))
+            ->merge(new StringNullableArray(['bar', 'baz']));
     }
 }